+++ src/main/java/com/example/delivery/common/exception/code/ErrorCode.java
@@ -26,9 +26,11 @@ public enum ErrorCode {
     NOT_FOUND_CART(HttpStatus.NOT_FOUND, "해당 장바구니를 찾을 수 없습니다."),
     UNAVAILABLE_MENU(HttpStatus.BAD_REQUEST, "해당 메뉴는 주문하실 수 없습니다."),
     CART_PERMISSION_DENIED(HttpStatus.FORBIDDEN, "해당 장바구니에 대한 접근 권한이 없습니다."),
+    CART_MODIFICATION_NOT_ALLOWED(HttpStatus.BAD_REQUEST, "해당 장바구니는 수정할 수 없습니다."),
     CANNOT_ADD_DIFFERENT_STORE_MENU(HttpStatus.BAD_REQUEST, "같은 가게 메뉴만 담으실 수 있습니다."),
     CANNOT_ADD_MENU_FROM_UNOWNED_STORE(HttpStatus.BAD_REQUEST, "자신 가게 메뉴만 담으실 수 있습니다."),
     NOT_FOUND_MENU_OPTION(HttpStatus.NOT_FOUND, "해당 옵션를 찾을 수 없습니다."),
+    INVALID_MENU_COUNT(HttpStatus.BAD_REQUEST, "메뉴는 1개 이상부터 주문이 가능합니다."),
 
     // 공통 에러
     INVALID_INPUT_VALUE(HttpStatus.BAD_REQUEST, "유효하지 않은 입력값입니다."),
